Initial implementation of Tooling API performance tests
This commit introduces a new kind of performance tests, aimed at measuring the performance of the Tooling API.
By extending `AbstractToolingApiCrossVersionPerformanceTest` a test can check the performance of the tooling
API for various scenarios. In that case, you can specify:

- a list of target versions
- an action to execute in a `GradleConnection` context

The list of target versions is used to tell which version of the Tooling API to use, but also to tell which
version of Gradle we're connecting to.

For now, we're only checking the total duration of the operation, and we are reusing the `CrossVersionResultsStore` to
store the results. We do *not* check the memory consumption yet (not our initial focus).